+++ src/main/java/com/chilun/deneng/controller/TaskForceController.java
@@ -6,8 +6,10 @@ import com.chilun.deneng.Response.BaseResponse;
 import com.chilun.deneng.Response.ResultCode;
 import com.chilun.deneng.pojo.Article;
 import com.chilun.deneng.pojo.TaskForce;
+import com.chilun.deneng.pojo.User;
 import com.chilun.deneng.service.IArticleService;
 import com.chilun.deneng.service.ITaskForceService;
+import com.chilun.deneng.tools.auth.AuthUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -27,6 +29,9 @@ public class TaskForceController {
     //增删改查
     @Autowired
     ITaskForceService service;
+
+    @Autowired
+    AuthUtil authUtil;
 
     @GetMapping
     public BaseResponse queryTaskForceById(@RequestParam int id) {
@@ -54,21 +59,35 @@ public class TaskForceController {
     }
 
     @PostMapping
-    public BaseResponse addTaskForce(@RequestBody TaskForce taskForce) {
+    public BaseResponse addTaskForce(@RequestBody TaskForce taskForce,
+                                     @SessionAttribute(name = "user", required = false) User user,
+                                     @CookieValue(name = "JWT", required = false) String jwt) {
+        if (!authUtil.isManager(user, jwt)) return new BaseResponse("非管理员", ResultCode.UN_AUTHOR);
         boolean save = service.save(taskForce);
         if (save) return new BaseResponse(null, ResultCode.SUCCESS);
         return new BaseResponse(null, ResultCode.FAILURE);
     }
 
     @PutMapping
-    public BaseResponse updateTaskForce(@RequestBody TaskForce taskForce) {
+    public BaseResponse updateTaskForce(@RequestBody TaskForce taskForce,
+                                        @SessionAttribute(name = "user", required = false) User user,
+                                        @CookieValue(name = "JWT", required = false) String jwt) {
+        TaskForce byId = service.getById(taskForce.getId());
+        if(byId==null){
+            return new BaseResponse("修改的任务组不存在", ResultCode.FAILURE);
+        }
+        if (!authUtil.isManager(user, jwt) && !authUtil.isWantedUser(user, jwt, byId.getLeaderId()))
+            return new BaseResponse("非管理员或任务组负责人", ResultCode.UN_AUTHOR);
         boolean update = service.updateById(taskForce);
         if (update) return new BaseResponse(null, ResultCode.SUCCESS);
         return new BaseResponse(null, ResultCode.FAILURE);
     }
 
     @DeleteMapping
-    public BaseResponse deleteTaskForce(@RequestParam int id) {
+    public BaseResponse deleteTaskForce(@RequestParam int id,
+                                        @SessionAttribute(name = "user", required = false) User user,
+                                        @CookieValue(name = "JWT", required = false) String jwt) {
+        if (!authUtil.isManager(user, jwt)) return new BaseResponse("非管理员", ResultCode.UN_AUTHOR);
         boolean remove = service.removeById(id);
         if (remove) return new BaseResponse(null, ResultCode.SUCCESS);
         else return new BaseResponse(null, ResultCode.FAILURE);
